feat: 수익률 계산 후 출력

+++ src/main/java/domain/Lottos.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class Lottos {
 
     final List<Lotto> lottos;
-    public Lottos(int issueCount){
+    public Lottos(long issueCount){
         List<Lotto> lottos = new ArrayList<>();
 
         for(int i=0; i<issueCount; i++){
